requirements.txt
annotated-types==0.7.0
anyio==4.9.0
contourpy==1.3.2
cycler==0.12.1
dnspython==2.7.0
exceptiongroup==1.3.0
fastapi==0.116.0
fastapi-utils==0.8.0
fonttools==4.58.5
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psutil==5.9.8
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.13.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
schedule==1.2.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
